fix(webapp/admin): Move collection path component into common/ for future re-use in Dashboard and question caching
