+++ frontend/src/metabase/browse/components/utils.tsx
@@ -1,36 +1,10 @@
 import { t } from "ttag";
 
-import type { CollectionEssentials, SearchResult } from "metabase-types/api";
+import { getCollectionPathString } from "metabase/common/components/CollectionPath/utils";
+import type { SearchResult } from "metabase-types/api";
 import { SortDirection, type SortingOptions } from "metabase-types/api/sorting";
 
 import type { ModelResult } from "../types";
-import { getCollectionName } from "../utils";
-
-import { pathSeparatorChar } from "./constants";
-
-export const getBreadcrumbMaxWidths = (
-  collections: CollectionEssentials["effective_ancestors"],
-  totalUnitsOfWidthAvailable: number,
-  isPathEllipsified: boolean,
-) => {
-  if (!collections || collections.length < 2) {
-    return [];
-  }
-  const lengths = collections.map(
-    collection => getCollectionName(collection).length,
-  );
-  const ratio = lengths[0] / (lengths[0] + lengths[1]);
-  const firstWidth = Math.max(
-    Math.round(ratio * totalUnitsOfWidthAvailable),
-    25,
-  );
-  const secondWidth = totalUnitsOfWidthAvailable - firstWidth;
-  const padding = isPathEllipsified ? "2rem" : "1rem";
-  return [
-    `calc(${firstWidth}cqw - ${padding})`,
-    `calc(${secondWidth}cqw - ${padding})`,
-  ];
-};
 
 export const isModel = (item: SearchResult) => item.model === "dataset";
 
@@ -40,16 +14,6 @@ export const getModelDescription = (item: SearchResult) => {
   } else {
     return item.description;
   }
-};
-
-export const getCollectionPathString = (collection: CollectionEssentials) => {
-  const ancestors: CollectionEssentials[] =
-    collection.effective_ancestors || [];
-  const collections = ancestors.concat(collection);
-  const pathString = collections
-    .map(coll => getCollectionName(coll))
-    .join(` ${pathSeparatorChar} `);
-  return pathString;
 };
 
 const getValueForSorting = (
