+++ frontend/src/metabase/browse/utils.ts
@@ -1,30 +1,10 @@
 import type { Dispatch, SetStateAction } from "react";
-import { t } from "ttag";
 import _ from "underscore";
 
-import {
-  canonicalCollectionId,
-  coerceCollectionId,
-  isRootCollection,
-} from "metabase/collections/utils";
 import { entityForObject } from "metabase/lib/schema";
 import type { IconName } from "metabase/ui";
-import type { CollectionEssentials } from "metabase-types/api";
 
 import type { FilterableModel } from "./types";
-
-export const getCollectionName = (collection: CollectionEssentials) => {
-  if (isRootCollection(collection)) {
-    return t`Our analytics`;
-  }
-  return collection?.name || t`Untitled collection`;
-};
-
-/** The root collection's id might be null or 'root' in different contexts.
- * Use 'root' instead of null, for the sake of sorting */
-export const getCollectionIdForSorting = (collection: CollectionEssentials) => {
-  return coerceCollectionId(canonicalCollectionId(collection.id));
-};
 
 export type AvailableModelFilters = Record<
   string,
