+++ frontend/src/metabase/common/components/CollectionPath/CollectionPath.styled.tsx
