+++ frontend/src/metabase/common/components/CollectionPath/CollectionPath.unit.spec.tsx
@@ -2,14 +2,11 @@ import { renderWithProviders, screen } from "__support__/ui";
 import type { Collection } from "metabase-types/api";
 import { createMockCollection } from "metabase-types/api/mocks";
 
-import { CollectionBreadcrumbsWithTooltip } from "./CollectionBreadcrumbsWithTooltip";
+import { CollectionPath } from "./CollectionPath";
 
 const setup = (collection: Collection) => {
   return renderWithProviders(
-    <CollectionBreadcrumbsWithTooltip
-      collection={collection}
-      containerName="Container"
-    />,
+    <CollectionPath collection={collection} containerName="Container" />,
   );
 };
 const collectionAlpha = createMockCollection({ id: 99, name: "Alpha" });
@@ -24,7 +21,7 @@ const collectionCharlie = createMockCollection({
   effective_ancestors: [collectionAlpha, collectionBeta],
 });
 
-describe("CollectionBreadcrumbsWithTooltip", () => {
+describe("CollectionPath", () => {
   it("should show a single collection", () => {
     setup(collectionAlpha);
     expect(screen.getByLabelText("Alpha")).toBeInTheDocument();
