+++ frontend/src/metabase/common/components/CollectionPath/PathSeparator.tsx
