+++ frontend/src/metabase/common/components/CollectionPath/constants.tsx
