+++ frontend/src/metabase/common/components/types.tsx
@@ -1,0 +1,5 @@
+import type { MutableRefObject } from "react";
+
+export type RefProp<RefValue> = {
+  ref: MutableRefObject<RefValue> | ((value: RefValue) => void);
+};
